Don't output error stack on test runner exit
If the test runner fails then funkie creates an empty error to throw, which then gets logged out. We don't want to do that because the error contains no useful information and the test runner output should be sufficient to debug.

Keep the error logging only for runtime code errors and not errors in the test runner process.